Add score

+++ test.h
@@ -6,7 +6,7 @@
 /*   By: abittel <abittel@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/10 14:58:25 by abittel           #+#    #+#             */
-/*   Updated: 2022/01/12 13:19:05 by abittel          ###   ########.fr       */
+/*   Updated: 2022/01/13 15:33:14 by abittel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,18 +26,18 @@ typedef struct s_data
 	int			des[8];
 	t_joueur	*players;
 	int			pikomino[16];
-}	t_data;
-void	print_table (t_data *data, int *score, int player);
-void	print_des (t_data *data, int nbr_des, int *valeurs);
-int		get_pts(int val);
-void	print_screen (int	**screen);
-void	print_str_in_screen (int**screen, char *obj, int x, int y);
-void	print_in_screen (int**screen, int **obj, int x, int y);
-int		**get_pikomino(int val, int pts);
-int		**get_blank_screen (int x, int y);
-int		get_yes_no(const char *str);
-void	clearScreen(void);
-int		ft_strlen(char *str);
-int		get_lst_pikomino (t_data *data, int player);
+}			t_data;
+void		print_table (t_data *data, int *score, int player);
+void		print_des (t_data *data, int nbr_des, int *valeurs);
+int			get_pts(int val);
+void		print_screen (long long	**screen);
+void		print_str_in_screen (long long**screen, char *obj, int x, int y);
+void		print_in_screen (long long **screen, long long **obj, int x, int y);
+long long	**get_pikomino(int val, int pts);
+long long	**get_blank_screen (int x, int y);
+int			get_yes_no(const char *str);
+void		clearScreen(void);
+int			ft_strlen(char *str);
+int			get_lst_pikomino (t_data *data, int player);
 #endif
